backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
redis>=4.6.0
boto3>=1.28.38
azure-identity>=1.13.0
python-multipart>=0.0.6
sse-starlette>=1.8.2
